Narrow Linux kernel symbol CRCs to 32 bits
These values are 32 bits wide.

PiperOrigin-RevId: 486145011

+++ crc.h
@@ -26,7 +26,7 @@
 namespace stg {
 
 struct CRC {
-  uint64_t number;
+  uint32_t number;
 };
 
 inline bool operator==(CRC crc1, CRC crc2) {
